requirements.txt
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
pandas>=2.0
torch>=2.0
torchaudio>=2.0
transformers>=4.35
datasets>=2.14
praatio>=6.0
soundfile>=0.12
pyyaml>=6.0
tqdm>=4.65
